+++ tests/unit/bot/crawlers/test_crawlers_init.py
@@ -1,5 +1,0 @@
-from ekrhizoc.bot.crawlers import crawlers
-
-
-def test_commands():
-    assert [C().name for C in crawlers] == ["universal-bfs"]