+++ tests/unit/bot/crawlers/test_universal_bfs.py
@@ -1,1 +1,0 @@
-# TODO: Add universal bfs crawler tests